// app/types/api/about.ts
interface About {
    id: number;
    created_at: string;
    updated_at: string;
    title: string;
    subtitle : string;
    photo: string;
    description: string;
    years_number: string;
    years_text: string;
    projects_number: string;
    projects_text: string;
    frameworks_number: string;
    frameworks_text: string;
    button_text: string
    button_icon: string;
    button_link:string;
    user_id: number;
    
  }
  
  export type { About };
  