// app/types/api/navbar.ts
export interface NavBar {
    id: number;
    section_one_text: string;
    section_one_link: string;
    section_two_text: string;
    section_two_link: string;
    section_three_text: string;
    section_three_link: string;
    section_four_text: string;
    section_four_link: string;
    section_five_text: string;
    section_five_link: string;
    created_at: string;
    updated_at: string;
}